+++ src/constants.ts
@@ -1,4 +1,7 @@
-import { RelationshipClass } from '@jupiterone/integration-sdk-core';
+import {
+  RelationshipClass,
+  RelationshipDirection,
+} from '@jupiterone/integration-sdk-core';
 
 export const SetDataKeys = {
   ACCOUNT_ENTITY: 'ACCOUNT_ENTITY',
@@ -6,18 +9,20 @@ export const SetDataKeys = {
 
 export const StepIds = {
   FETCH_ACCOUNT: 'fetch-account',
+  FETCH_PROJECTS: 'fetch-projects',
   FETCH_FINDINGS: 'fetch-findings',
+  FETCH_USERS: 'fetch-users',
 };
 
 export const Entities = {
   CVE: {
     _type: 'cve',
-    _class: 'Vulnerability',
+    _class: ['Vulnerability'],
     resourceName: 'CVE',
   },
   CWE: {
     _type: 'cwe',
-    _class: 'Weakness',
+    _class: ['Weakness'],
     resourceName: 'CWE',
   },
   SNYK_ACCOUNT: {
@@ -27,8 +32,18 @@ export const Entities = {
   },
   SNYK_FINDING: {
     _type: 'snyk_finding',
-    _class: 'Finding',
+    _class: ['Finding'],
     resourceName: 'Snyk Issue',
+  },
+  PROJECT: {
+    _type: 'snyk_project',
+    _class: ['Project'],
+    resourceName: 'Snyk Project',
+  },
+  USER: {
+    _type: 'snyk_user',
+    _class: ['User'],
+    resourceName: 'Snyk User',
   },
 };
 
@@ -50,5 +65,27 @@ export const Relationships = {
     sourceType: Entities.SNYK_ACCOUNT._type,
     _class: RelationshipClass.IDENTIFIED,
     targetType: Entities.SNYK_FINDING._type,
+  },
+  ACCOUNT_PROJECT: {
+    _type: 'snyk_account_has_project',
+    sourceType: Entities.SNYK_ACCOUNT._type,
+    _class: RelationshipClass.HAS,
+    targetType: Entities.PROJECT._type,
+  },
+  ACCOUNT_USER: {
+    _type: 'snyk_account_has_user',
+    sourceType: Entities.SNYK_ACCOUNT._type,
+    _class: RelationshipClass.HAS,
+    targetType: Entities.USER._type,
+  },
+};
+
+export const mappedRelationships = {
+  PROJECT_REPO: {
+    _type: 'snyk_project_scans_coderepo',
+    sourceType: Entities.PROJECT._type,
+    _class: RelationshipClass.SCANS,
+    targetType: 'CodeRepo',
+    direction: RelationshipDirection.FORWARD,
   },
 };