+++ src/index.ts
@@ -1,12 +1,22 @@
 import { IntegrationInvocationConfig } from '@jupiterone/integration-sdk-core';
 
-import instanceConfigFields from './instanceConfigFields';
-import { steps } from './steps/fetchFindings';
-import { IntegrationConfig } from './types';
-import validateInvocation from './validateInvocation';
+import {
+  IntegrationConfig,
+  instanceConfigFields,
+  validateInvocation,
+} from './config';
+import { steps as accountSteps } from './steps/account';
+import { steps as projectSteps } from './steps/projects';
+import { steps as findingSteps } from './steps/fetchFindings';
+import { steps as userSteps } from './steps/users';
 
 export const invocationConfig: IntegrationInvocationConfig<IntegrationConfig> = {
   instanceConfigFields,
   validateInvocation,
-  integrationSteps: steps,
+  integrationSteps: [
+    ...accountSteps,
+    ...projectSteps,
+    ...findingSteps,
+    ...userSteps,
+  ],
 };