+++ src/steps/users/index.test.ts
@@ -1,0 +1,11 @@
+import { createStepCollectionTest } from '../../../test/recording';
+import { StepIds } from '../../constants';
+
+test(
+  'fetch-users',
+  createStepCollectionTest({
+    directoryName: __dirname,
+    recordingName: 'fetch-users',
+    stepId: StepIds.FETCH_USERS,
+  }),
+);