+++ CHANGELOG.md
@@ -8,6 +8,46 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Support for ingesting the following **new** entities:
+
+  | Resources                 | Entity `_type` | Entity `_class` |
+  | ------------------------- | -------------- | --------------- |
+  | Snyk Project              | `snyk_project` | `Project`       |
+  | Snyk Organization Members | `snyk_user`    | `User`          |
+
+- Added support for ingesting the following **new** relationships:
+
+  | Source         | class   | Target         |
+  | -------------- | ------- | -------------- |
+  | `snyk_account` | **HAS** | `snyk_project` |
+  | `snyk_account` | **HAS** | `snyk_user`    |
+
+- Added support for ingesting the following **new** mapped relationships:
+
+  | Source         | class     | Target     |
+  | -------------- | --------- | ---------- |
+  | `snyk_project` | **SCANS** | `CodeRepo` |
+
+- New properties added to entities:
+
+  | Entity         | Properties               |
+  | -------------- | ------------------------ |
+  | `snyk_account` | `id`                     |
+  | `snyk_account` | `name`                   |
+  | `snyk_account` | `function`               |
+  | `snyk_finding` | `isPinnable`             |
+  | `snyk_finding` | `isFixable`              |
+  | `snyk_finding` | `isPartiallyFixable`     |
+  | `snyk_finding` | `name`                   |
+  | `snyk_finding` | `violatedPolicyPublicId` |
+
+### Fixed
+
+- Fix value of `numericSeverity` by converting the string severity value to a
+  number
+
 ## [2.1.2] - 2022-03-08
 
 ### Fixed